+++ core/Test/Tasty/CmdLine.hs
@@ -1,5 +1,5 @@
 -- | Parsing options supplied on the command line
-{-# LANGUAGE CPP, ScopedTypeVariables, DeriveDataTypeable #-}
+{-# LANGUAGE CPP, ScopedTypeVariables, DeriveDataTypeable, RecordWildCards #-}
 module Test.Tasty.CmdLine
   ( optionParser
   , suiteOptions
@@ -7,13 +7,18 @@ module Test.Tasty.CmdLine
   , defaultMainWithIngredients
   ) where
 
-import Options.Applicative
 import Data.Monoid
 import Data.Proxy
-import Data.Foldable (foldMap)
+import Data.Tagged
+import Data.Either (lefts, rights)
+import Data.Maybe
 import Prelude  -- Silence AMP and FTP import warnings
 import System.Exit
 import System.IO
+import System.Environment (getArgs)
+import qualified System.Console.GetOpt as GetOpt
+import Text.Printf
+import Control.Monad (when)
 
 -- We install handlers only on UNIX (obviously) and on GHC >= 7.6.
 -- GHC 7.4 lacks mkWeakThreadId (see #181), and this is not important
@@ -33,19 +38,44 @@ import System.Mem.Weak (deRefWeak)
 import Test.Tasty.Core
 import Test.Tasty.Ingredients
 import Test.Tasty.Options
+import Test.Tasty.Options.CLParser
+import Test.Tasty.Options.Core
 import Test.Tasty.Options.Env
-import Test.Tasty.Runners.Reducers
-
 
 -- | Generate a command line parser from a list of option descriptions
-optionParser :: [OptionDescription] -> Parser OptionSet
-optionParser = getApp . foldMap toSet where
-  toSet :: OptionDescription -> Ap Parser OptionSet
-  toSet (Option (Proxy :: Proxy v)) = Ap $
-    (singleOption <$> (optionCLParser :: Parser v)) <|> pure mempty
+optionParser :: [OptionDescription] -> [GetOpt.OptDescr (Either String OptionSet)]
+optionParser = map $ \(Option (Proxy :: Proxy v)) ->
+  let
+    Tagged name = optionName :: Tagged v String
+    Tagged help = optionHelp :: Tagged v String
+    CLParser{..} = optionCLParser :: CLParser v
+    metavar_ = fromMaybe "ARG" clMetavar
+    parseOpt :: String -> Either String OptionSet
+    parseOpt str =
+      case parseValue str of
+        Just (val :: v) -> Right $ singleOption val
+        Nothing -> Left $ printf "Invalid value for option %s: %s"
+          name (show str)
+    argDescr :: GetOpt.ArgDescr (Either String OptionSet)
+    argDescr =
+      case clOnValue of
+        Nothing -> GetOpt.ReqArg parseOpt metavar_
+        Just val ->
+          GetOpt.OptArg
+            (maybe (Right $ singleOption val) parseOpt)
+            metavar_
+  in
+    GetOpt.Option
+      clShort
+      [name]
+      argDescr
+      help
 
 -- | The command line parser for the test suite
-suiteOptionParser :: [Ingredient] -> TestTree -> Parser OptionSet
+suiteOptionParser
+  :: [Ingredient]
+  -> TestTree ->
+  [GetOpt.OptDescr (Either String OptionSet)]
 suiteOptionParser ins tree = optionParser $ suiteOptions ins tree
 
 -- | Parse the command line arguments and run the tests using the provided
@@ -57,15 +87,29 @@ suiteOptionParser ins tree = optionParser $ suiteOptions ins tree
 defaultMainWithIngredients :: [Ingredient] -> TestTree -> IO ()
 defaultMainWithIngredients ins testTree = do
   installSignalHandlers
-  cmdlineOpts <- execParser $
-    info (helper <*> suiteOptionParser ins testTree)
-    ( fullDesc <>
-      header "Mmm... tasty test suite"
-    )
+  args <- getArgs
+  let
+    opt_descrs = suiteOptionParser ins testTree
+    (opts_or_errs, _, errs0) = GetOpt.getOpt GetOpt.RequireOrder opt_descrs args
+    errs1 = errs0 ++ lefts opts_or_errs
+    usage = GetOpt.usageInfo "Mmm... tasty test suite\n\nAvailable options:" opt_descrs
+  cmdlineOpts <-
+    if null errs1
+      then return . mconcat . rights $ opts_or_errs
+      else do
+        mapM_ (hPutStrLn stderr) errs1
+        hPutStrLn stderr "" -- add an extra empty line after the errors
+        hPutStrLn stderr usage
+        exitFailure
 
   envOpts <- suiteEnvOptions ins testTree
 
   let opts = envOpts <> cmdlineOpts
+      Help help = lookupOption opts
+
+  when help $ do
+    putStrLn usage
+    exitSuccess
 
   case tryIngredients ins opts testTree of
     Nothing -> do